fix(drawer): Fixed a bug where the drawer would not close when the menu was clicked
---
There was a bug where the drawer would not close automatically when a menu item was clicked. This was especially bad for the user experience on small screen widths such as mobile.

In this change, we added an onClick event to the <Box> component that contains the menu, and fixed it so that `isSmDown` (when the screen width is narrow) dispatches `setDrawerOpen(false)`. This ensures that the drawer will close as expected when any item in the drawer is clicked.