refactor (ToDo): Organizing display variables in ToDoTableRow
We've reorganized display variables in the ToDoTableRow component to improve code readability.

The logic for displaying a hyphen `-` as a default value for null or undefined values has been moved from within the JSX to the variable declaration at the top of the component.
This simplifies the JSX and consolidates the logic in one place, improving maintainability.