feat: ToDoテーブルから期日列を削除しパッケージ列を追加 (#865)
* feat: Display package name instead of due date in ToDo list

The ToDo list now shows the package name associated with a task instead of the due date.

This change provides more relevant information at a glance for vulnerability management tasks, as the package name is often crucial for identifying and addressing vulnerabilities. The due date information was less frequently utilized in this view.

* refactor (ToDo): Organizing display variables in ToDoTableRow

We've reorganized display variables in the ToDoTableRow component to improve code readability.

The logic for displaying a hyphen `-` as a default value for null or undefined values has been moved from within the JSX to the variable declaration at the top of the component.
This simplifies the JSX and consolidates the logic in one place, improving maintainability.