feat: Unify the Pagination component
### Changes

The `Pagination` component in `ToDoCardView.jsx` has been **assigned to a variable named `paginationComponent` and unified**. This allows the same component to be reused in two different places.

### Reason

Previously, the same `Pagination` component code was **duplicated** in two different `Box` containers. The purpose of this change is to eliminate code duplication and centralize the component's logic.

This makes future maintenance easier, as any changes to the appearance or behavior of the `Pagination` component will **only need to be made in one place**. This is an improvement that aligns with the DRY (Don't Repeat Yourself) principle.